refactor(lib::podcast): extract "Podcast" structs and impls into own module
